# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Your API Description",
        "title": "Your API Name",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/logs": {
            "get": {
                "description": "Returns a list of all logs from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get all logs",
                "responses": {
                    "200": {
                        "description": "List of logs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backendT_internal_database_repository.Log"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/logs/filtered": {
            "get": {
                "description": "Returns filtered logs based on method, response type and time range. Requires limit, offset and timeRange parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get filtered logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "HTTP method to filter by",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Response status code to filter by",
                        "name": "response",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time range (e.g. '-1 hour', '-24 hours', '-7 days'). Required parameter.",
                        "name": "timeRange",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination. Required parameter.",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination. Required parameter.",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backendT_internal_database_repository.LogsGetBasicViewWithOffsetLimitAdvancedRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/logs/paginated": {
            "get": {
                "description": "Returns a paginated list of logs with basic view",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logs"
                ],
                "summary": "Get paginated logs without filters",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backendT_internal_database_repository.LogsGetBasicViewWithOffsetLimitRow"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "description": "Returns a list of all posts from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get all posts",
                "responses": {
                    "200": {
                        "description": "List of posts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backendT_internal_database_repository.Post"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new post in the database. Expects a JSON body with the required fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "parameters": [
                    {
                        "description": "New post payload",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backendT_internal_database_repository.PostsCreateParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created post",
                        "schema": {
                            "$ref": "#/definitions/backendT_internal_database_repository.Post"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/posts/id/{id}": {
            "get": {
                "description": "Fetches a single post by numeric ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get post by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found post",
                        "schema": {
                            "$ref": "#/definitions/backendT_internal_database_repository.Post"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/posts/userid/{userid}": {
            "get": {
                "description": "Fetches a single post by its user ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get post by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found post",
                        "schema": {
                            "$ref": "#/definitions/backendT_internal_database_repository.Post"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid user ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backendT_internal_database_repository.Log": {
            "type": "object",
            "properties": {
                "bytes_in": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "bytes_out": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "error": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "host": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "id": {
                    "type": "integer"
                },
                "latency": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "latency_human": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "method": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "remote_ip": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "request_id": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "status": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "timestamp": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "uri": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "user_agent": {
                    "$ref": "#/definitions/sql.NullString"
                }
            }
        },
        "backendT_internal_database_repository.LogsGetBasicViewWithOffsetLimitAdvancedRow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "method": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "path": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "response": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "response_time": {
                    "$ref": "#/definitions/sql.NullString"
                }
            }
        },
        "backendT_internal_database_repository.LogsGetBasicViewWithOffsetLimitRow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "method": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "path": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "response": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "response_time": {
                    "$ref": "#/definitions/sql.NullString"
                }
            }
        },
        "backendT_internal_database_repository.Post": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "backendT_internal_database_repository.PostsCreateParams": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer",
                    "format": "int64"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}